backend/pyproject.toml
[project]
name = "web-dev-26"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.11.1",
    "dj-database-url==2.3.0",
    "django==5.2.13",
    "django-cors-headers==4.9.0",
    "djangorestframework==3.17.1",
    "djangorestframework-simplejwt==5.3.1",
    "factory-boy>=3.3.3",
    "gunicorn==21.2.0",
    "psycopg2-binary==2.9.9",
    "pyjwt==2.8.0",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "sqlparse==0.5.5",
    "stripe==9.0.0",
    "whitenoise==6.8.2",
]
